+++ NEWS.md
@@ -4,6 +4,7 @@
 * `bartlettTests()` gains formula and lm methods
 * `colDevs()` gains a `group.var` argument; examples extended
 * Added `uniStats()` for an anova version of `glance.mlm()`
+* Corrected documentation of `covEllipses()` to more accurately describe recycling of arguments.
 
 ## Version 1.7.2 (2024-11-25)